+++ RollbarNotifier/Sources/RollbarNotifier/RollbarLogger.m
@@ -182,6 +182,11 @@ static BOOL isNetworkReachable = YES;
 - (void)logCrashReport:(NSString *)crashReport {
 
     RollbarConfig *config = self.configuration;
+    
+    if (YES == [self shouldSkipReporting:config.loggingOptions.crashLevel]) {
+        return;
+    }
+    
     RollbarPayload *payload = [self buildRollbarPayloadWithLevel:config.loggingOptions.crashLevel
                                                          message:nil
                                                        exception:nil
@@ -189,38 +194,51 @@ static BOOL isNetworkReachable = YES;
                                                            extra:nil
                                                      crashReport:crashReport
                                                          context:nil];
-    if (payload) {
-        [self queuePayload:payload.jsonFriendlyData];
-    }
+    [self report:payload];
 }
 
 - (void)log:(RollbarLevel)level
     message:(NSString *)message
+       data:(NSDictionary<NSString *, id> *)data
+    context:(NSString *)context {
+    
+    if (YES == [self shouldSkipReporting:level]) {
+        return;
+    }
+
+    RollbarConfig *config = self.configuration;
+    
+    RollbarPayload *payload = [self buildRollbarPayloadWithLevel:level
+                                                         message:message
+                                                       exception:nil
+                                                           error:nil
+                                                           extra:data
+                                                     crashReport:nil
+                                                         context:context
+                               ];
+    [self report:payload];
+}
+
+- (void)log:(RollbarLevel)level
   exception:(NSException *)exception
        data:(NSDictionary<NSString *, id> *)data
     context:(NSString *)context {
     
-    if (!self.configuration.developerOptions.enabled) {
+    if (YES == [self shouldSkipReporting:level]) {
         return;
     }
     
     RollbarConfig *config = self.configuration;
     
-    if (level < config.loggingOptions.logLevel) {
-        return;
-    }
-
     RollbarPayload *payload = [self buildRollbarPayloadWithLevel:level
-                                                         message:message
+                                                         message:nil
                                                        exception:exception
                                                            error:nil
                                                            extra:data
                                                      crashReport:nil
                                                          context:context
                                ];
-    if (payload) {
-        [self queuePayload:payload.jsonFriendlyData];
-    }
+    [self report:payload];
 }
 
 - (void)log:(RollbarLevel)level
@@ -228,15 +246,11 @@ static BOOL isNetworkReachable = YES;
        data:(NSDictionary<NSString *, id> *)data
     context:(NSString *)context {
 
-    if (!self.configuration.developerOptions.enabled) {
+    if (YES == [self shouldSkipReporting:level]) {
         return;
     }
     
     RollbarConfig *config = self.configuration;
-    
-    if (level < config.loggingOptions.logLevel) {
-        return;
-    }
     
     RollbarPayload *payload = [self buildRollbarPayloadWithLevel:level
                                                          message:nil
@@ -246,6 +260,25 @@ static BOOL isNetworkReachable = YES;
                                                      crashReport:nil
                                                          context:context
                                ];
+    [self report:payload];
+}
+
+- (BOOL)shouldSkipReporting:(RollbarLevel)level {
+    
+    RollbarConfig *config = self.configuration;
+    
+    if (!config.developerOptions.enabled) {
+        return YES;
+    }
+    
+    if (level < config.loggingOptions.logLevel) {
+        return YES;
+    }
+    
+    return NO;
+}
+
+- (void)report:(RollbarPayload *)payload {
     if (payload) {
         [self queuePayload:payload.jsonFriendlyData];
     }