+++ RollbarNotifier/Sources/RollbarNotifier/include/Rollbar.h
@@ -28,19 +28,32 @@
 + (void)updateConfiguration:(RollbarConfig *)configuration;
 + (void)reapplyConfiguration;
 
-#pragma mark - New logging methods
+#pragma mark - Logging methods
+
++ (void)logCrashReport:(NSString*)crashReport;
 
 + (void)log:(RollbarLevel)level
     message:(NSString *)message;
 + (void)log:(RollbarLevel)level
-    message:(NSString *)message
   exception:(NSException *)exception;
 + (void)log:(RollbarLevel)level
+      error:(NSError *)error;
+
++ (void)log:(RollbarLevel)level
     message:(NSString *)message
+       data:(NSDictionary<NSString *, id> *)data;
++ (void)log:(RollbarLevel)level
   exception:(NSException *)exception
        data:(NSDictionary<NSString *, id> *)data;
 + (void)log:(RollbarLevel)level
+      error:(NSError *)error
+       data:(NSDictionary<NSString *, id> *)data;
+
++ (void)log:(RollbarLevel)level
     message:(NSString *)message
+       data:(NSDictionary<NSString *, id> *)data
+    context:(NSString *)context;
++ (void)log:(RollbarLevel)level
   exception:(NSException *)exception
        data:(NSDictionary<NSString *, id> *)data
     context:(NSString *)context;
@@ -49,62 +62,69 @@
        data:(NSDictionary<NSString *, id> *)data
     context:(NSString *)context;
 
-+ (void)debug:(NSString *)message;
-+ (void)debug:(NSString *)message
-    exception:(NSException *)exception;
-+ (void)debug:(NSString *)message
-    exception:(NSException *)exception
-         data:(NSDictionary<NSString *, id> *)data;
-+ (void)debug:(NSString *)message
-    exception:(NSException *)exception
-         data:(NSDictionary<NSString *, id> *)data
-      context:(NSString *)context;
+#pragma mark - Convenience logging methods
 
-+ (void)info:(NSString *)message;
-+ (void)info:(NSString *)message
-   exception:(NSException *)exception;
-+ (void)info:(NSString *)message
-   exception:(NSException *)exception
-        data:(NSDictionary<NSString *, id> *)data;
-+ (void)info:(NSString *)message
-   exception:(NSException *)exception
-        data:(NSDictionary<NSString *, id> *)data
-     context:(NSString *)context;
++ (void)debugMessage:(NSString *)message;
++ (void)debugException:(NSException *)exception;
++ (void)debugError:(NSError *)error;
 
-+ (void)warning:(NSString *)message;
-+ (void)warning:(NSString *)message
-      exception:(NSException *)exception;
-+ (void)warning:(NSString *)message
-      exception:(NSException *)exception
-           data:(NSDictionary<NSString *, id> *)data;
-+ (void)warning:(NSString *)message
-      exception:(NSException *)exception
-           data:(NSDictionary<NSString *, id> *)data
-        context:(NSString *)context;
++ (void)debugMessage:(NSString *)message data:(NSDictionary<NSString *, id> *)data;
++ (void)debugException:(NSException *)exception data:(NSDictionary<NSString *, id> *)data;
++ (void)debugError:(NSError *)error data:(NSDictionary<NSString *, id> *)data;
 
-+ (void)error:(NSString *)message;
-+ (void)error:(NSString *)message
-    exception:(NSException *)exception;
-+ (void)error:(NSString *)message
-    exception:(NSException *)exception
-         data:(NSDictionary<NSString *, id> *)data;
-+ (void)error:(NSString *)message
-    exception:(NSException *)exception
-         data:(NSDictionary<NSString *, id> *)data
-      context:(NSString *)context;
++ (void)debugMessage:(NSString *)message data:(NSDictionary<NSString *, id> *)data context:(NSString *)context;
++ (void)debugException:(NSException *)exception data:(NSDictionary<NSString *, id> *)data context:(NSString *)context;
++ (void)debugError:(NSError *)error data:(NSDictionary<NSString *, id> *)data context:(NSString *)context;
 
-+ (void)critical:(NSString *)message;
-+ (void)critical:(NSString *)message
-       exception:(NSException *)exception;
-+ (void)critical:(NSString *)message
-       exception:(NSException *)exception
-            data:(NSDictionary<NSString *, id> *)data;
-+ (void)critical:(NSString *)message
-       exception:(NSException *)exception
-            data:(NSDictionary<NSString *, id> *)data
-         context:(NSString *)context;
 
-+ (void)logCrashReport:(NSString*)crashReport;
++ (void)infoMessage:(NSString *)message;
++ (void)infoException:(NSException *)exception;
++ (void)infoError:(NSError *)error;
+
++ (void)infoMessage:(NSString *)message data:(NSDictionary<NSString *, id> *)data;
++ (void)infoException:(NSException *)exception data:(NSDictionary<NSString *, id> *)data;
++ (void)infoError:(NSError *)error data:(NSDictionary<NSString *, id> *)data;
+
++ (void)infoMessage:(NSString *)message data:(NSDictionary<NSString *, id> *)data context:(NSString *)context;
++ (void)infoException:(NSException *)exception data:(NSDictionary<NSString *, id> *)data context:(NSString *)context;
++ (void)infoError:(NSError *)error data:(NSDictionary<NSString *, id> *)data context:(NSString *)context;
+
++ (void)warningMessage:(NSString *)message;
++ (void)warningException:(NSException *)exception;
++ (void)warningError:(NSError *)error;
+
++ (void)warningMessage:(NSString *)message data:(NSDictionary<NSString *, id> *)data;
++ (void)warningException:(NSException *)exception data:(NSDictionary<NSString *, id> *)data;
++ (void)warningError:(NSError *)error data:(NSDictionary<NSString *, id> *)data;
+
++ (void)warningMessage:(NSString *)message data:(NSDictionary<NSString *, id> *)data context:(NSString *)context;
++ (void)warningException:(NSException *)exception data:(NSDictionary<NSString *, id> *)data context:(NSString *)context;
++ (void)warningError:(NSError *)error data:(NSDictionary<NSString *, id> *)data context:(NSString *)context;
+
++ (void)errorMessage:(NSString *)message;
++ (void)errorException:(NSException *)exception;
++ (void)errorError:(NSError *)error;
+
++ (void)errorMessage:(NSString *)message data:(NSDictionary<NSString *, id> *)data;
++ (void)errorException:(NSException *)exception data:(NSDictionary<NSString *, id> *)data;
++ (void)errorError:(NSError *)error data:(NSDictionary<NSString *, id> *)data;
+
++ (void)errorMessage:(NSString *)message data:(NSDictionary<NSString *, id> *)data context:(NSString *)context;
++ (void)errorException:(NSException *)exception data:(NSDictionary<NSString *, id> *)data context:(NSString *)context;
++ (void)errorError:(NSError *)error data:(NSDictionary<NSString *, id> *)data context:(NSString *)context;
+
++ (void)criticalMessage:(NSString *)message;
++ (void)criticalException:(NSException *)exception;
++ (void)criticalError:(NSError *)error;
+
++ (void)criticalMessage:(NSString *)message data:(NSDictionary<NSString *, id> *)data;
++ (void)criticalException:(NSException *)exception data:(NSDictionary<NSString *, id> *)data;
++ (void)criticalError:(NSError *)error data:(NSDictionary<NSString *, id> *)data;
+
++ (void)criticalMessage:(NSString *)message data:(NSDictionary<NSString *, id> *)data context:(NSString *)context;
++ (void)criticalException:(NSException *)exception data:(NSDictionary<NSString *, id> *)data context:(NSString *)context;
++ (void)criticalError:(NSError *)error data:(NSDictionary<NSString *, id> *)data context:(NSString *)context;
+
 
 #pragma mark - Send manually constructed JSON payload